Added option for prepending "namespace::" to functions
This more closely mirrors the output provided by IDA since it does not
treat namespaces as a separate part of the function name, at least in
the context of binexport.

Functions belonging to different classes that all implement the same
interface can now be properly identified in BinDiff and name matching is
more useful.