Use pre-generated API stubs to avoid requiring a Binary Ninja install
This change includes a script that extracts all C API functions from the
Binary Ninja API and creates empty stub functions for them. This is later
compiled into an API compatible library that is used to link against. This
avoids having to use install a copy of Binary Ninja on the machine that builds
BinExport.

If an installation of Binary Ninja is available, then BinExport can be built
using that, too.

Note that the script can be replaced with a small Clang tool to make it more
robust should that become necessary.

PiperOrigin-RevId: 336267078
Change-Id: Id3b3db1b0f0212a3c9318ca8f7deb100dc81bdc3