Add a tool to perform config file servicing
This tool allows the installers and package install scripts to configure the
per-machine config file, avoiding the need to run shell scripts as part of the
install. On Windows, this is particularly useful, as Windows Installer custom
actions are notoriously brittle when implemented as a script (which would need
to be a JScript/VBScript!).

The tool also functions to set up per-user configuration, creating symlinks to
the disassembler plugins in the user's home folder. For example for IDA Pro on
Linux, it will create symlinks in `~/.idapro/plugins`. This enables IDA Pro to
load the plugins without requiring the user to specify its installation
directory.

PiperOrigin-RevId: 344547919
Change-Id: Ib4eff79e22c2efc44a1dac47c57a0615d488135e