Work around MSVC being too aggressive in LTO and LTCG
When compiling with Visual Studio on Windows, when the CMake target property
`INTERPROCEDURAL_OPTIMIZATION` is enabled, Abseil's flag parsing breaks in
subtle ways:

```
bindiff.exe --primary a --secondary b
```

Will result in both `absl::GetFlag(FLAGS_primary)` and
`absl::GetFlag(FLAGS_secondary)` being set to the string `1`. In the
command-line above, the argument values `a` and `b` are actually treated as
positional arguments in this case (and BinDiff will error with "Extra
arguments").

This change disables `INTERPROCEDURAL_OPTIMIZATION` for the command-line
differ.

Other changes to make command-line handling more robust:
- `--nologo` => `--logo`, inverting the default. Abseil will still understand
  `--nologo`, but will error at `--nonologo`.
- The Java UI will now always specify command-line arguments with an equals
  sign: `bindiff --primary=a --secondary=b`

PiperOrigin-RevId: 370447655
Change-Id: I84b9a68e36069bd8912df0d00a342064a6328b73