Refactor logging and use plain `msg()` in IDA
This prevents a dead-lock with the log executor that surfaced when building
with IDA SDK 7.6. Additionally, the log executor was not really necessary at
all, as `msg()`/`vmsg()` are marked as being thread-safe.

Also remove `executable` and `executable64` from BinDiff config. Those were
never actually used by the Java UI in released versions and we do not support
using BinDiff with `idat`/`idat64` anyways.

This change also actually implements file logging in the plugins if requested.
Before, this could only be requested on the command-line.

Other chages:
- Suppress MSVC compiler warnings in IDA SDK, disable some globally

PiperOrigin-RevId: 374132575
Change-Id: I74c31664a5f2aad1c1bda907521b6028c8a8c8ab