macOS: Implement a "first stage" UI launcher to behave well under TCC
This new first stage launcher just selects and runs the native JVM launcher
binary for the correct architecture (AArch64 or x86-64) in order to present
BinDiff as Universal Binary.

While this can be (and was!) implemented in a simple shell script, this
does not play well with macOS' Transparency, Consent, and Control (TCC)
mechanism.

For example, using a shell script launcher will lead to "Operation not
permitted" errors instead of a permission prompt, when Java code tries
to access files in protected folders (like "Desktop", "Documents", etc.).
This is because the originating binary (i.e. "/bin/sh" or similar) will
not invoke any UX and receive an error unless "Full Disk Access" was
granted manually in System Preferences.

PiperOrigin-RevId: 376157999
Change-Id: I79c4dccf7756b30fe99c9b9e46c11e153ac1f380