UI: Remove "Window Fix" hack
On Linux, if the `DESKTOP_SESSSION` environment variable is set, ZyLib tries
to access obsolete internal JDK classes via reflection and crashes.

These classes were finally removed in Java 16 and the "window fix", the
`GuiHelper` class was implementing is also no longer necessary, as both the
Window Managers and the Java runtime have been fixed in the meantime.

This issue was not found during testing, as Chrome Remote Desktop does not seem
to set `DESKTOP_SESSION`. Bummer.

PiperOrigin-RevId: 377859970
Change-Id: Iee98297078587ff1f389bfa7a495ddc58f3564eb