Add a partial data reference type
In Aarch64, data references are usually worked out with an `ADRP` instruction and then a later `ADD`.  The `ADD` may follow immediately, or it may be separated by potentially multiple other instructions.  The ADRP can be considered to be a "partial" reference, with the final reference being known when the `ADD` instruction is encountered.

RISC-V has a similar system, with `lui` or `auipc` followed by `addi` or using the base register directly in `lw` or `sw`.

Adding a partial reference type means we can track when this has occurred.  Potentially the base register, as loaded by `ADRP`, could also be used in multiple subsequent `ADD` instructions.

PiperOrigin-RevId: 473777970
Change-Id: If42ec9cc018288c35ab8503864a891f2b79a8721